[level 1] Title: 정수 제곱근 판별, Time: 0.03 ms, Memory: 75.6 MB -BaekjoonHub

+++ 프로그래머스/lv1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -4,8 +4,8 @@ class Solution {
         long answer = 0;
         
         
-        if(Math.pow((int)Math.sqrt(n) ,2) == n ){
-            answer = (long) Math.pow(Math.sqrt(n)+1 ,2);
+        if(Math.pow((int)Math.sqrt(n),2) == n ){
+            answer = (long)Math.pow((int)Math.sqrt(n)+1,2);
         }else{
             answer = -1;
         }
